// src/app/time-recording-core/models/enum.ts
export enum StatusType {
    Fail = 0,
    Success = 1
}

export enum TimeRecordingCloseInfo {
    ExitByUser = 'ExitByUser',
    ExitWithSaveSuccess = 'ExitWithSaveSuccess',
    Minimize = 'Minimize',
    DeleteTR = 'DeleteTR'
}
export enum StopStartInfo {
    Start = 'START',
    Stop = 'STOP',
}
export enum TimeRecordAddingType {
    GeneralTimeSave = 'GeneralTimeSave',
    FloatingTimeSave = 'FloatingTimeSave',
    FloatingTimeUnpostToPost = 'FloatingTimeUnpostToPost'
}
// export enum FormSaveType {
//     Save = 'Save',
//     Submit = 'Submit',
// }